CMS7-8493: change styling of the current document listing
The hippo data table has become one tree instead of three separate trees (one for the header, body, and footer). The body is scrollable with use of CSS only (see document-listing.html for the proof-of-concept implementation). JavaScript is only needed to keep the width of the thead and tbody in sync; the width of all columns is kept in sync by CSS.

All data table-related CSS has moved to the central .less files.

AFAICS tables.js is not used anywhere anymore and can be removed.