CMS7-8509: combine type and state icon in one column
Also refactored the constructors of HippoIcon into static factory methods to improve code readability. 

The state icon has to be rendered as an inline svg to be able to style the individual svg paths via CSS 
(a linked svg sprite icon generates a shadow dom that cannot be styled individually anymore).