HSTTWO-3154 implemented support for creating all available links for some document within a single mount or across all mounts. The difference with createAll and createAllCanonicalLinks is that createAll returns also *within* a single mount possibly multiple links (thus not only the canonical). Also now, support for link resolving has been added for documents that are linked via a component item via JcrPath or DocumentLink annotation. If a document can both be resolved by a component item its JcrPath or DocumentLink and a sitemap its relative contentpath, the normal precedence rules for the best sitemap item are used. (for example sitemap item for the "/" root is preferred above "/some/deeper/path"). We still need to enhance this part to make sure that a current typical pitfall happens:
Suppose you have a news article that can be resolved via sitemap item '/news/**.html' and which can be resolved via sitemap item '/home' because that sitemap item happens to point to a page component, which has a component item with a JcrPath/DocumentLink parameter pointing as well to the news article. Requesting an HstLink for that news article then results in the /home sitemap item, as this is the shorter one (which has precedence). However, typically, the homepage shows for example only the intro of the news article and links further to the real URL for the news article. However, this link would now result again in the homepage. This is undesirable most likely in general. We need to be able to inject some extra (customizable?) behavior  

Note : The support for component item linked documents in linkrewriting is only available for a mount if the hst:site node for that mount has hst:componentlinkrewritingsupported = true. By default, support for component item linked documents link rewriting is switched off (for now)