CMS7-8513: style left border of document editor
The border of .hippo-browser-documents-small-files (the div around the collapsed document list) has been removed, and replaced with a left border on the document list head and body+foot and the editor top and body. This made it possible to use a slightly darker border color for the head/top part.