CMS7-8711: prevent interleaving of file system changes and processing of these changes to ensure stable watcher tests
The watcher tests used to wait for event bus activity before verifying behavior, but the start of file system change
processing was not synchronized. It could therefore happen that the file watcher already started observing
changes (e.g. when the file poller thread fired and started scanning) while the test modified the file system.
As a result, the observed changes by the watcher did not always match the test expectations. To prevent 
any interleaving, the web files watcher now has additional callbacks to notify its listeners that it started and 
stopped watching a set of changes, which are used by the test to prevent thread interleaving.