CHANNELMGR-114: keep component properties of non-saved variants, replace hst_internal_variant usage with component properties preview
If a component variant was saved, the forms of all other variants used to be reinitialized, wiping any changes they had. 
We now only create a new properties form for the saved variant, and reuse the forms of all other variants. As a result,
the changes in these non-saved forms are kept, and can still be previewed and saved afterwards.

The dirty state of each form is now calculated 'manually' by comparing the initial value and current value of each record.
The form's dirty states could not be used because Ext.form.Field#isDirty only works after a field has been rendered.
When a variant is saved, reused dirty forms in non-active tabs are not rendered yet and effectively lost their dirty state. 
Checking the records in the copied stores fixed this.