CHANNELMGR-135 Fix disabling of save button
The save button should be disabled if there are no pending
changes or if a variant is created that is initially identical
to the original. There were two problems:
1. Hidden fields were also checked for equality
Fixed by skipping the hidden fields.
2. Number fields containing int values were compared against
initial values that are always strings (because that is how
the hst currently returns them).
Fixed by always comparing string values. For the initial value
coercing to a String is in principle not needed, because it is
already a string, but it shows the intent of the comparison more
clearly.