CHANNELMGR-153 Extend cms api IFramePanel
The IFramePanel (and MessageBus) have been moved to cms api but
the isValidSession function was not (because it is too channelmanager
specific). Since we do need this function in the channel manager I
have extended Hippo.IFramePanel and (re)implemented the function by
using the Hippo.IFramePanel.getCookies() function. I also deleted the
MessageBus.js because it has moved to cms-api too.