CMS7-9018: include external usage-statistics.js script in the CMS
The external script is loaded via Ajax to prevent blocking the CMS UI. An 'async' script tag was considered but not used because an unreachable async script causes an annoying spinner in the browser tab. The 'timeout' property of the ajax call ensures that jQuery invokes the error callback when the script cannot be found.