CHANNELMGR-195: show breadcrumb when channel manager is hard-refreshed
With a hard refresh the channel manager is already activated, so the 'channel-manager-activated' event was not published and hence the breadcrumb would not be shown. The 'isActivated' state is now passed from the server to the client. If the channel manager is already activated, the client shows the breadcrumb, else hides it.