CMS7-9244: let CodeMirror do scrolling, use flexbox to make edited script full height, and divide height between executed script and feedback log.
Because CodeMirror now scrolls instead of its container element, both browser zoom and very long lines are  rendered correctly.
  
Flexbox does resize the container of the script correctly, but does not set a height. The height of CodeMirror is therefore kept in sync with its container's height using jQuery. When there's logging feedback the available height is split between the script and the feedback panel. The script height is then set to half of the total height of the script plus feedback panel, so flexbox gives the feedback panel the other half. 

The window resize event fires on both window resizes as browser zoom, which is exactly what we need.