CHANNELMGR-24: also save string version of date as initial value, re-enable manual editing
Manual editing allows a user to reset a date to its default value, or to clear it in case 
no default value is defined. 

The 'keyup' event has been replaced by 'valid', so the value is only updated when a field 
contains a valid value. This prevents illegal dates from being saved to the record.
 
The 'specialkey' callback has to explicitly check for invalid empty date values, otherwise
a date with a default value can still be saved as "no date" (i.e. an empty string) when a 
user edits a date manually, clears the field and very quickly types ENTER. In that case the
'specialkey' event happens before the 'valid' event, and since a date field thinks an empty
date is also valid it will save the empty date instead of resetting it to the default value.