CHANNELMGR-297: include the deleted variant IDs in the 'variantsDeleted' event
This way the properties panel knows which variant tabs can be reused or not.
It also solves the edge case where an experiment on Default A, B, and C is
completed and only Default B is kept. The backend will then promote Default B
to the new Default. The variantsDeleted event will now indicate that Default A
and Default C have been deleted, so the Default tab (which is Default A) won't
be reused and will be rendered again. That's good, because it will fetch the data
from Default B that way, and show the correct variant configuration.