CHANNELMGR-307: ping all site webapps to keep their session alive
The channel manager used to ping the webapp that contained the viewed/edited channel while the Channels perspective was active. As a result, the session of other webapps (e.g. a separate intranet webapp) would time out if it was not viewed within the session timeout of the webapp. Also other perspectives that call REST services in the site could get an invalidated session at unexpected times (e.g. when doing a POST) because the channel manager ping was disabled while the channel manager was not active.
 
The solution for now is to let the channel manager ping all webapps all the time so all their sessions are kept alive. Because the keepalive is a GET, the SSO redirect handshake will be triggered automatically by a ping whenever the session does time out. The keepalive call returns the maximum inactive interval of the session, which is used to optimize the keepalive call frequency (one minute less than the max. interval, with a minimum of 20 seconds to remain backwards compatible). Failed pings (e.g. because a webapp cannot be reached anymore) are silently ignored, and assumed to be one-off errors.

The initial pings at startup are necessary to a) retrieve the max inactive interval, and b) to reset previous sessions that belong to another user that was recently active on the same machine (because a logout currently does not actively invalidate site sessions).