CHANNELMGR-282: adjust renderPathInfo when user browses channel
The first edit of a channel creates its -preview configuration. In that case the location 
of the iframe used to be reset to the initial location when the channel was viewed, 
i.e. the home page of that channel. This was caused by the renderPathInfo in the page 
container, which was not kept up-to-date when a user browsed the channel. This has now
been fixed: the render path is updated after each iframe page load, and the latest render path
info is provided in the browseTo call after the -preview configuration has been created.