CMS-9114: add 'stay logged in' option to login screen
Checking the option disables the active logout for a user. The selected value is stored in a cookie. The option is unchecked by default. The checkbox is only shown for the login screen of the CMS application.

The 'session.timeout.minutes' config option has been moved to /hippo:configuration/hippo:frontend/settings/max.inactive.interval, and accepts the Wicket duration syntax.