REPO-1429 support newly added nodetypes can be found by already logged in session
The SearchAfterNodeTypeChangeTest is a test that failed without the fix in this issue. The fix is that there is a global counter that keeps track of everytime a nodetype change is found. After a nodetype change, the authorisation query will be recreated (and thus 'nodetype expansion' in the query will also include the new nodetype). Note I had to add a hippoecm-extension.xml to the test jar to add user 'editor' to the group 'editor' and likewise for 'author'. Namely the users do exist, but they are in 'config module' not added to the groups. Since we should not add them by default (apparently), I have added them by default in the test setup only. All tests still pass (at least for repo but I expect downstream projects as well)