CHANNELMGR-465 restart form validation when same component is reopened
The client-side form validation ensures that the Save button is
enabled when a form field is changed. The validation is stopped
when the component properties window is hidden (i.e. closed) to improve
client-side performance (the validation loop runs continuously).
Because we skipped all logic when the same component was opened
again, the validation was not started again, so the Save button never
became enabled again.