CHANNELMGR-476 add runtime dependency of cm-frontend to cm-frontend-ng
The artifact 'hippo-addon-channel-manager-frontend' is already
included as a dependency of the cms war in all existing projects.
Making it depend 'hippo-addon-channel-manager-frontend-ng' ensures that
the NG code is pulled in automatically in all existing projects without
making any changes to the cms pom.xml file.

It can be a 'runtime' dependency since the frontend does not compile
against the frontend-ng code.