CHANNELMGR-470 Add ScrollService for scrolling the page while dragging
Wire up the ScrollService in the DragDropService and use the DragDropService.dragging state to decide if scrolling can start.
Currently the ScrollService scrolls only up and down (using velocity animations). It attaches a mouse-enter and mouse-leave event handler to the container element and uses these trigger to initiate or stop a scroll on the page.
it only attaches event listeners to browsers that lack native scroll support while dragging.
The duration of a scroll animation is between 500 and 1500 ms based on the distance to scroll, although this is a very simple impl. We might want to slow it down a bit to 500-2000 ms.