CHANNELMGR-528 enable drag and drop within iframe in IE11
The MouseEvent() constructor does not work in IE11 (despite what the
Mozilla docs claim), so we have to use the deprecated initMouseEvent()
method. IE11 also needs an MS-specific name ('MSPointerDown' instead of
'mousedown') for the generated event, otherwise it won't be picked up
by Dragula.

Generation of the mousedown event moved to the DomService so all
browser-specific logic is done in there.

With this in, clicking a component no longer worked. It turned out that
IE11 calls the per-component mouseup handler *after* Dragula called the
onDrop callback (which differed from what all other browers do). So in
IE11 we were no longer dragging when the mouseup handler was called and
hence the component properties were never shown when a component was
clicked.

Clicking a component was fixed by registering only a single mouseup
event handler on the component that is being dragged. That event handler
is removed again when dragging stops. This is both simpler (only one
event handler needed and only while dragging) and also works in IE11.
We use a namespaced mouseup event name so jQuery's off() only removes
the event handler we registered.