CHANNELMGR-512 notify app when channel changed
Saving component parameters and deleting component configuration
variants changes a channel outside the Angular app. These actions now
trigger a reload of the channel store, which will update the channel
overview. When that's done a 'channel-changed' event with the latest
channel data is sent to the app, which will also update the
'changedBySet' of the channel and hence show/hide the 'Changes'
toolbar item.