CHANNELMGR-523 Scroll page up and down when dragging an item from the components catalog
Use the dom-autoscroller module [1] to scroll a page up and down when dragging an item out of bounds. Since it is only available as a NPM package, we have to 'map' the complete dependency graph; for every require() statement a related file should be mapped into the systemjs configuration. Apart from this, Karma should also be configured to serve these files.

 [1] https://github.com/hollowdoor/dom_autoscroller