CHANNELMGR-571 show dragged mirror outside iframe
The mirror used to be cut off because it was rendered inside the
channel-iframe-scroll element.

This could not be fixed by setting overflow-x property of the scroll
element to 'visible' because the 'auto' value is needed to render a
scrollbar when the contents is too wide (e.g. the site is not
responsive).

The mirror wrapper is now rendered outside the channel-iframe-scroll
element, and positioned on top of the iframe whenever dragging
starts.