CHANNELMGR-580 set view-as selection to the page render variant
The current render variant of the page is stored server-side, and
included in the meta data when a page is rendered. The selected
variant in the 'view-as' dropdown should always be initialized to
this render variant.

By passing the render variant as an attribute to the
relevance-view-as directive, the view-as selection is updated
automatically when the page is (re)loaded.