CHANNELMGR-611 add 'runtime' checks if component can be moved
Components should not be moved in or into disabled containers.
Disabled containers are not registered when the page is loaded,
but can still turn disabled when another user changed them, the
current user tries to modify them (which fails), and then the
containers are re-rendered. Since a container cannot be 'removed'
from Dragula, the 'moves' and 'accepts' methods are used to check
that the source and/or target container is indeed enabled.

Also refactored the unit test a little. Spying on a container
object is not possible before they have been created, so several
existing tests that assumed disabled containers were not doing so
at all. Instead additional comment data (like being disabled) can
now be set before loading the iframe fixture.