CHANNELMGR-603 Don't unsubscribe from CmsService event in callback
The message bus cannot handle an unsubscribe of an event during a
callback for the same event. That happened when a pick was canceled
and a subsequent pick was not.

We now subscribe to the 'picked' event whenever a controller for a
JcrPath property is created, and unsubscribe from it when the
controller is destroyed again.