CHANNELMGR-594 use horizontal drag direction for HST.Span containers
Requires that dragula's 'direction' option can be function so the
drag direction can be specified per container instead of only once
for all containers on a page. 'Direction as function' is still an
open pull request (https://github.com/bevacqua/dragula/pull/336),
so we forked dragula into the onehippo GitHub repository
(https://github.com/onehippo/dragula) to get it in. Once the pull
request is accepted and released the fork can be removed again.