CHANNELMGR-337 indicate progress while channel or page are loading
The main toolbar now contains a indeterminate linear progress
indicator that is shown when either the channel is not loaded yet,
or the page is not loaded yet.

A 'clear-channel' event is to clear the existing channel. This
will hide the iframe. Once the channel is loaded the iframe is
shown again. Once the page inside the iframe is loaded the
progress indicator is disabled.

To prevent odd interaction while the channel is loading, all items
in the main toolbar are either disabled (sidenav, device switcher,
edit/view) or hidden (view as, changes).