CHANNELMGR-638 (substantial) rework after review
The current code is lacking a solid design to distribute the task of editing a menu clearly over the SiteMenuService and the MenuEditorController. I made some fixes to make save and cancel actually work, and started some improvements towards a clearer separation of concerns. Due to timing constraints, I guess this is sufficient for the very short term.