CMS-10069 support for fallback to lower micro versions for translations artifacts
Maven l10n plugin include locales mojo should fall back on latest available micro version of translation artifacts.
This is now supported by checking the version z.y.(x-1) if z.y.x is missing. If x is not a number but for example
3-SNAPSHOT or 3-cmng-psp1, the (x-1) will result in '2' because that can be the earlier released version. The
fallback is done until micro version '0'. If that version can still not be found, a message is logged that the
locale artifact can not be found.
If the micro version is '03' or '03-SNAPSHOT', the fallback micro version is '2' without the preceding '0'. '03' is namely not our semantic versioning scheme but we  should (and do) use '3' instead