CHANNELMGR-723 Remove superfluous hippoTree.isEmpty checks
The surrounding button already has an ng-if for !hippoTree.isEmpty,
so the button's icon is only rendered when the tree item is not
empty. No need to repeat the check in the ng-show expression of the
icon name.

Also used a ternary expression for the icon name to prevent
additional span elements (results in a smaller DOM).