HSTTWO-3554 Honoring a TTL when that is set by an HstComponent
When an HstComponent sets an 'Expires' header, make sure the cached page
gets a timeToLive until the page expires. Also note that there was a small
bug in the ESIPageInfoScanningValve resulting in a uncacheable response still
being cached. Hence I changed HstCacheEhCacheImpl#put to first inspect whether the
element to cache is cacheable : If not, put null.

Note that HstPageInfo#isNoCachePresentOrExpiresImmediately was broken wrt the
Expires header because it tried to parse the value as long while the value is actually
in a GMT date format. I fixed this by correctly parsing the date Expires header

Also now use ehcache.getCacheConfiguration().getMaxEntriesLocalDisk() because #getMaxElementsOnDisk
was deprecated and javadoc said getMaxEntriesLocalDisk should be used for non-clustered caches