CHANNELMGR-759 Fix height calculation for certain sites
- if a website does not explicitly set it's height to 100% on the
body/html element, the overlay sync service fails to detect the
correct height. To fix this we also take the
offsetHeight/scrollHeight of the HTML element into account (using
jQuery's height function).
This does break the UI (the grey canvas is revealed when scrolling
down) so we need to set the height on the sheet element and only make
the scrollX element higher if the horizontal scrollbar is visible.