ESSENTIALS-991 Update LICENSE file
Cherry picked from 7842572

Delete from LICENSE
- angular-tablesort.js
The angular-tablesort.js library is a transitive dependency of hippo-theme.
However, in the dashboard.war binary the the tablesort library is not present.
Therefore ther reference to the tablesort library should not be present.

Add to LICENSE
- bootstrap.min.js
- prettify.js
- jquery-1.8.0.min.js
- jquery.ba-bbq.min.js
- backbone-min.js
- handlebars-1.0.0.js
- jquery.wiggle.min.js
- underscore-min.js
- highlight.7.3.pack.js
because they are packaged in the war.