HIPPLUG-1329 [Forward port 3.1]  make sure every worker has its own urlset
The Urlset object should have to think about synchronization. It is the code
that uses the Urlset that shouldn't expect the Urlset to be thread-safe (just
like a HashSet isn't thread-safe. It make much more sense to not give all the
workers a reference to the same urlset object but all have their own

A seperate note is that I really think the worker should not extend from
Thread and a lot of the complex logic can be replaced by a simple executer service

(cherry picked from commit 031750fa2dfb14e8a9fd3671839d7b0df2c7021b)