CHANNELMGR-824 Container-specific drag direction without Dragula fork
The drag direction ('vertical' or 'horizontal') can be different per
container. We used to use a fork of the Dragula library that
contained the 'direction-as-function' pull request. Yet instead we
can keep a reference to the 'options' object passed to Dragula, and
update the 'direction' property in it when needed. This makes
Dragula use the updated value instead. The fork is therefore not
needed anymore.

The direction is updated when a drag starts and when the mouse moves
over a container.