CMS-10320 Translate wizard buttons
The setLabel call in AjaxWizardButton did not do anything (that
label is only used in validator messages). Instead the label model
is now given to the constructor. The wizard labels were also not
translated anywhere, so added a .properties file for them. Finally,
the hardcoded 'value' in the <input> elements overrode the
translation in the model, so removed those. Also changed the input
elements to type 'button' since we are submitted via Ajax anyway.