HSTTWO-3728 Move canDeleteChannel to composermode response
The data of the separate /features call can also be returned in
the composermode call, which (once we get rid of /features) saves
a REST call.

Also moved the canManageChanges logic to here. For now both
canDeleteChannel and canManageChanges do not check the actual
permissions of the current user. That will be fixed later.