CHANNELMGR-794 Composermode returns canManageChanges/canDeleteChannel
Instead of the separate /features call the canDeleteChannel is now
returned by the /composermode call. This will save a REST call in
the future (once /features is gone entirely).

The canManageChanges permission is now also returned by the
composermode call, so the JCR permissions logic can be removed from
the channel manager.