CHANNELMGR-796 Remove unused listener for 'mountChanged' event
The old channel manager used the 'mountChanged' event to signal
the overview that the detail view switched to another channel
(e.g. because someone clicked a link to another channel). This
event is no longer used in the new channel manager. Instead, it
uses the 'switch-channel' event which is handled by the
ChannelEditor itself.