ARCHE-509 Correct the log4j to account for the MdcOrJndiPropertyFilter
Since we use a MdcOrJndiPropertyFilter that matches on 'cms' or 'site' logs, we must
change

<root>
  <level value="warn"/>
  <appender-ref ref="cms"/>
</root>

into

<root>
  <level value="warn"/>
  <appender-ref ref="cms"/>
  <appender-ref ref="site"/>
</root>

to make sure also logging to 'site' is ending up in the hippo-site.log file.

And in all loggers, we can remove the 'additivity' (it doesn't matter, it should
match only one logger via MdcOrJndiPropertyFilter any way) *and* we can remove all
the <appender-ref ref="xyz"/> except for the audit logging which goes to a specific
log file without MdcOrJndiPropertyFilter filter.