HIPPLUG-1357 Remove the unused (since broken) SitemapFeedBasedOnHstSitemap and SiteMapBasedOnHstSiteMapResource sitemap generators
As a result, quite some other code was unused, which in turn triggered other code to be possible to be removed.

Note I also do a content delete for

/hst:hst/hst:configurations/hst:default/hst:components/forge-sitemap-based-on-hst-configuration-feed

since this component refers to SitemapFeedBasedOnHstSitemap.

I've tested that after a restart (with existing repository), that 'forge-sitemap-based-on-hst-configuration-feed'
component is gone, basic sitemap and NewsSitemapFeed still work as expected