CMS-10444 Only execute workflow in setMode() when available
The setMode() implementation obtains or commits an editable instance
of the document when the mode is set to 'edit'. However, obtaining
or committing an editable instance is not always possible (e.g. when
a publication request is pending).

We now check whether a workflow method is available (via the workflow
hints) before actually invoking it. The actual mode is only changed
when the workflow call succeeds to keep the behavior the same
(super.setMode() was not called when the workflow method failed).