CMS-10348 Move update logic of members to group members list view
The group members list view is thereby responsible for updating
its own model.

Pass the list of permissions and members directly as model so Wicket
takes care of the serializable check. Since the used ArrayList is
actually serializable this saves copying the whole list just to mark
it as Serializable.

Also reduced code clutter:
- Reuse a single groupModel
- Use Model.of() where possible
- Remove the use of PropertyModel where possible
- Remove serialVersionUID