HSTTWO-3780 Improved fluent api wrt HstQueryBuilder
The HstQueryBuilder had the following methods:

    public static FilterBuilder filter(String fieldName) {
        FieldFilterBuilder filterBuilder = new FieldFilterBuilder(fieldName);
        return filterBuilder;
    }

    public static FilterBuilder and(FilterBuilder ... filterBuilders) {
        AndFilterBuilder filterBuilder = new AndFilterBuilder(filterBuilders);
        return filterBuilder;
    }

    public static FilterBuilder or(FilterBuilder ... filterBuilders) {
        OrFilterBuilder filterBuilder = new OrFilterBuilder(filterBuilders);
        return filterBuilder;
    }

As a result, it was tempting to create a query as follows:

HstQuery hstQueryInFluent = HstQueryBuilder.create()
                .scopes(baseContentBean)
                .filter("myhippoproject:customid").equalTo("123")
                .queryBuilder().build();

This compiled perfectly fine and is tempting to do, however #build() results in an NPE. The
above is not a valid construct, but this is quite unclear from the fluent api. Moving the methods
that return a FilterBuilder *to* the FilterBuilder and out of HstQueryBuilder makes sure the
compiler doesn't put you on the wrong foot. Note I also renamed
org.hippoecm.hst.content.beans.query.builder.FilterBuilder#filter to
org.hippoecm.hst.content.beans.query.builder.FilterBuilder#filterBuilder because I think that
is more explicit and less confusing when you use it in the fluent api : Now you get

HstQuery hstQueryInFluent = HstQueryBuilder.create()
                .scopes(baseContentBean)
                .filter(filterBuilder("myhippoproject:customid").equalTo("123"))
                .build();

which I think is more clear than

HstQuery hstQueryInFluent = HstQueryBuilder.create()
                .scopes(baseContentBean)
                .filter(filter("myhippoproject:customid").equalTo("123"))
                .build();

Note I added some new unit test method placeholders which I still need to do