HSTTWO-3835 Better compile time guidance of the fluent query api
Instead of returning a ConstraintBuilder for #and, #or and #constraint method, and
instead of being able to invoke many methods (recursively), have the fluent api
better guide the user : Instead, for #and and #or, return directly a Constraint object on
which the only method possible is #negate. The method #constraint returns a
ConstraintBuilder. To satisfy the compiler, developers *have* to invoke one of the
ConstraintBuilder methods (#equalTo, #contains, #exists, etc). After that method,
a Constraint is returned, and only #negate can be invoked.

Now the api is compile time much more informative.

Lastly, I moved the TestHstQueryBuilder to  a different package to guarantee that
access modifiers are correct for end users(quite some methods are
package private / protected).