CHANNELMGR-936 Always load a page by settings the 'src' attribute
The code used to keep track of the loaded path in the iframe. When
the render path to load was equal to the already loaded path the
iframe's location was reloaded.

This does not work when the 'src' attribute of the iframe differs
from the URL of loaded page (e.g. when the user navigated inside the
iframe). In that case a reload of the iframe's location would
actually load the original location in the 'src' attribute.

To fix this, we always update the 'src' attribute when explicitly
loading a new page.

Also ensured 100% line coverage of the hippo iframe service.