CHANNELMGR-966 Remove session refresh invocations
After discussing the current use of session.refresh() with Ate and Ard, we concluded
that we can best document the session to be "invocation scoped", i.e. the session is
created before the invocation of a DocumentsService or DocumentTypesService method,
and avoid the session.refresh() calls, as they consume quite some performance.