CHANNELMGR-966 Remove TODO
These two constants are also defined in the CMS API. I considered moving them to the repository and sharing them, but the repository is currently unaware of these "resource bundle groups" and terminology, and changing the API would be a lot of hassle. I also don't want to depend on the CMS (API) in the content service, so I figured duplicating this here would be the least bad option.