HSTTWO-3854 Only flush non-message-body contributions during redirect
Instead of not flushing components at all any more for a redirect, still flush them but
do not flush the message body (to avoid the response from possibly being committed
before the redirect takes place resulting in "Exception during sendRedirect.
java.lang.IllegalStateException: Cannot call sendRedirect() after the response has been committed")

This could already happen in version 10 (more likely in the channel mngr), but it was not so easily triggered. In
version 11, the problem is triggered much easier. Since there seems to be no reason to flush components before
redirecting, just remove that code

Also the comment in HstServletResponseState above the method getParentWriter was not accurate: The
response only become committed if the parent writer was the actual backing http servlet response writer *and*
on that writer #flush would be invoked. Hence removing the faulty comment